# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apollia OS Runtime API",
    "description": "Host driving contract for the Apollia OS sovereign agent runtime. Stable under /api/v1; breaking changes ship under /api/v2.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0-preview"
  },
  "paths": {
...
    "/api/v1/llm/recommend": {
      "get": {
        "tags": [
          "model_hub"
        ],
        "summary": "`GET /api/v1/llm/recommend?limit=...&n_ctx=...`",
        "description": "Ranks the models this machine should run: plans against the embedded\ngeneration table, resolves the survivors on HuggingFace, reads each\nfinalist's GGUF header over a range request, and orders what is left.\n\nAnswers `503` when the Hub cannot be reached, which is deliberately not the\nsame as an empty `200`. The catalogue is fetched live, so no network means\nno catalogue; reporting that as \"no models fit\" would be a different and\nfalse claim about the operator's machine.",
        "operationId": "recommend_models",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of recommendations",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "n_ctx",
            "in": "query",
            "description": "Context window the cache is sized against",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "hf_token",
            "in": "query",
            "description": "HuggingFace token for gated models",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Models ranked for this machine, best first"
          },
          "503": {
            "description": "HuggingFace unreachable, so there is no catalogue to rank",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/llm/registry/model/{org}/{repo}": {
      "get": {
        "tags": [
          "model_hub"
        ],
        "summary": "`GET /api/v1/llm/registry/model/:org/:repo?hf_token=...`",
        "operationId": "get_registry_model",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "description": "HuggingFace organization or user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "repo",
            "in": "path",
            "description": "HuggingFace repository name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hf_token",
            "in": "query",
            "description": "HuggingFace token for gated models",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model metadata and file list (HuggingFace model card)"
          },
          "403": {
            "description": "Model is gated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "Upstream HuggingFace error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/llm/registry/search": {
      "get": {
        "tags": [
          "model_hub"
        ],
        "summary": "`GET /api/v1/llm/registry/search?q=...&limit=...&sort=...`",
        "operationId": "search_registry",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort key",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hf_token",
            "in": "query",
            "description": "HuggingFace token for gated models",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching GGUF models (HuggingFace model cards)"
          },
          "502": {
            "description": "Upstream HuggingFace error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "description": "Canonical error body for every non-2xx response in the spec.\n\nMatches the wire shape of the per-module `ErrorResponse { error }` used\nacross the route handlers.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error description."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token read from `<data_dir>/api-token`, generated on first start and sent as `Authorization: Bearer <token>`. Enforced on the TCP listener by `TokenAuthLayer`, on every route except `POST /webhooks/{id}`, which authenticates its caller with an HMAC-SHA256 signature instead. Requests arriving on the Unix socket bypass the check: the socket file is mode 0600, so its permissions carry the same rule. Setting `[api] require_token = false` in `apollia.toml` removes the layer from the TCP listener altogether."
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}